Change the speech rate and volume handlers to take a floating-point multiplier for the default setting. (dm)
git-svn-id: svn://mielke.cc/main/brltty@1468 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/spk_driver.h
@@ -51,11 +51,11 @@ static void spk_close (void);
 #endif /* SPK_HAVE_TRACK */
 
 #ifdef SPK_HAVE_RATE
-  static void spk_rate (int setting);		/* mute speech */
+  static void spk_rate (float setting);		/* mute speech */
 #endif /* SPK_HAVE_RATE */
 
 #ifdef SPK_HAVE_VOLUME
-  static void spk_volume (int setting);		/* mute speech */
+  static void spk_volume (float setting);		/* mute speech */
 #endif /* SPK_HAVE_VOLUME */
 
 #ifdef SPKPARMS
